Moving classes into their own module
